fix: move the MAC address calculation to the script after loading the NAME variable
The MAC address of the containers is calculated based on the container's NAME. Due to recent changes in the order of loading configuration files, the MAC address was being calculated using an empty NAME string.
More details of the changes in #49

To fix this, the name_based_mac_addr function and its usage have been moved to the main script. This ensures that the variables are loaded in the correct order:

  1. Global config file - General variables declared in the repository.
  2. Project config file - Local variables specific to each project.
  3. Run script - Final configurations that depend on the global or project configurations being loaded beforehand.

This change prevents the error and ensures the MAC address is calculated correctly.

Closes #50